requirements.txt
aiogram==3.15.0
python-dotenv
yandex-cloud-ml-sdk

